All objects now have an associated id. This is to allow implementation of radio groupings, and other features, where two parts of the system may be linked purely by an id reference. Added `assign_id', `set_id', `update_internal_id_references' and `id'. Now inherits GB_SHARED_ID.
Added `all_children_recursive', `update_for_delete' and `delete' which is
to be called just before the object is deleted. Redefine in descendents
that need special processing for this.

Removed `xml_storage_required' as we always need to store the internal
data now, as the id will always be saved to xml.