`generate_code' now has the follow arguments:
	element: XML_ELEMENT; info: GB_GENERATED_INFO

previously it had these arguments:

	element: XML_ELEMENT; a_name, a_type: STRING; children_names ARRAYED_LIST [STRING]

The change was made to allow more information to be passed, without
having to keep adding new arguments ever time a new piece of information
was required by a certain type.

GB_GENERATED_INFO is a new class, specifically for this purpose, and
is part of a tree structure generated in a prepass stage of the
generation.

Added preconditions which did not exists before and improved comment.