Fixed bug when you did the following -
create dialog
dialog.show_relative_to_window (first_window)
dialog.hide
dialog.show


The dialog would then be shown relative again. We now inherit from
EV_DIALOG_IMP_COMMON twice, in order to get two versions of `show' which
we can use for this fix. This is not a clean solution, but seemed to be the
best of a bad set.