Fixed bug in `insert_i_th'. We now call the new item actions before we call `notify_change' to fix the following problem :-
If you have an EV_BOX with one or more items expansion disabled, and you
insert a widget before some or all of the items that are
`is_item_expanded', then the wrong widgets will not be expanded. This does
not always occur, as it depends on another resize event being triggered.

Now that `notify_change' is executed first, it recomputes the non expanded
children correctly, and then when `notify_change' is executed, the correct
list of expanded items is used.