Fixed `show_title_row' and `hide_title_row' which were adding `Lvs_single_Selction' when really it should have not been added, and not doing anything when it was removed.
This was always broken, but only recently was found, due to the change
in the default state not being in mutliple selection anymore. As the
widget is no longer in multiple selection by default, the first time
you call `hide_title_row' you experience the bug, whereas before, the
first time was ok.