Fixed two bugs :-
Adding a constant, permitted you to still press the add button again,
thereby ending up with two constants with the same name. This crashed
Build at a later date.

if no constant was selected, and you typed the name of an existing constant,
the "add" button was not disabled, again causing problems.