Fixed major bug which crashed EiffelBuild if you performed the following:
1. Converted an existing widget structure into a top level object.
2. Built one or more objects into this newly created top level object.
3. Undid the history to before the conversion of the top level object.
4. Re-did the history to the addition of the objects in the top level object.

Build crashed. This was becaue each time we call `execute' we have to
re-build the complete widget structure which was setting new ids to the
top level object. Other commands that relied on these ids crashed as no
objects with the original ids exist anymore. Now on the first execution
we store all ids and for each subsequent one, we restore the new objects
to have the original ids.