If I use EV_UNTITLED_DIALOG_IMP on GTK, when Ctrl + Tab dialog pop up, the dialog can't get focus. Even not working if call EV_UNTITLE_DIALOG.set_foucs just after dialog shown. If I use EV_WIDNOW_IMP, the Ctrl + Tab dialog will have focus just after it pop up as expected.
The changes on Windows are due to the changing of EV_SHADOW_DIALOG on Linux. EV_SHADW_DIALOG have to inherit from EV_WINDOW instead of EV_UNTITLED_DIALOG on Linux. So, on the Windows side, EV_SHADOW_DIALOG_IMP has to adapt the changing. There is no functional changing on Windows EV_SHADOW_DIALOG implementations.