Added local optimizations.
Changed inline_agent_table so that it is only created on demand instead of at creation time as most classes probably won't have an inline agent in them.  We also have the query has_inline_agents to avoid the creation of the table when calling the feature, this should be replaced when we have initializers.