This commit contains all changes necessary to build an EiffelStudio delivery which does not overwrite an official version of EiffelStudio with the same major/minor version number. The wix toolkit scripts and the source code currently will build an EiffelStudio 6.1 CDD Edition. This edition has its own registry entries including preferences and its own %application_data% folder.
The setup project is required to build the setup.dll required for building an .msi installer.

NOTE: The changes should only affect the windows version of EiffelStudio for the most part, and it has been tested on windows. I did not test it on Linux. If there should be problems with EiffelStudio builds on Linux due to this change, i'd be glad for a quick feedback!