Now replicated features are generated correctly, be it from conforming or non-conforming parents.
Now replicated features have delegated calls cast to their access class and not their written class.  This fixes non-conforming code generation as the static code is always generated in the access class and may possibly fix inheritance of unselected features as if the cast is made to the written class then polymorphically there is no path to a unselected routine, possibly meaning that calls to unselected routines from a descendent always call the selected version, this has to be further investigated.