Reworked the processor_tag_type creation a bit, to be easier to duplicate and also more correct implementation of the ordering.
Adding code to the feature-checker that corresponds to the
(*) and (X) operators from the SCOOP thesis.
Also, have implemented the is_controlled function from the thesis.

All of the above still needs to be tested, this will probably be
in the next commit, after I have seen if it is worthwhile to use the eweasel
testing framework for this task, otherwise the tests may be more ``offline''
until I find the best way to include them.