Fiddling around with the types hierarchy and finding out where the processor tag fits best, the version now kinda works. Further testing on the primitive types is needed to tell whether this has worked correctly (generics too).
Un-inherit the processor tag from the "Comparable" class, as it is not
a true ordering, where I need a more pure definition. In the meantime, I just
define my own "<" operator, and redefine "is_equal" to get my desired effect.

There is an out-of-source test file that I'm also working on, but I don't
want to include it here until I find the proper place to put it. No use in
muddying the waters yet.