Some more of the SCOOP generic rule is now implemented, but conditional as it won't compile the basecode because it is more restrictive.
A bug with the equality of the new scoop types has been fixed,
it was generating incorrect CECIL code and causing the 
C compiler to crash late in the compile.

Cleaned up the processor_tag_type a bit more, perhaps some more
cleanups can be done in cl_type_a.e

Implementing a first pass at `is_controlled' that is slightly deficient
but I think works in the simple cases that we'll start with.
Eventually more rules may be needed to more accurately decide when a
class is controlled, but for now I'm fairly sure we only admit
controlled classes, but reject a few too.