Adding a tests directory.
Currently the tests directory only contains a work-in-progress test
that I'm using to find typechecker bugs. 

To run the tests, just use the Makefile (although it is a little
unix-centric right now with the path-delimiters and usage of the
`touch' program).

Unfortunately it's all in one basic file, because I'm not sure how else
to effectively make tests for the compiler. I will try to work on making
the tests clear. 

As bugs are found the test-cases that produce them should be placed here.