Merging SCOOP into EVE.
The main bulk SCOOP work has been segregated into the scoopli and 
scoop2scoopli directories.

However, there is a nontrivial intersection in the type-checker.
These changes are mostly confined to the evaluted_type directory,
and also the ast_feature_checker_generator.

An extra degree is added conditioned on if the `separate' keyword
appears in the program text. Programs without this keyword
_should_ be unaffected by any SCOOP changes.

The parser was also updated, in conjunction with class_* so that
one can pass separate information from the AST to the internal
class representations.