Adding the modification of the ANY class.
The ANY class is modified to remove the export restrictions from
default_create. This is either done directly (as now) or also can 
be done with the scoop_any_roundtrip class.

There is an accompanying TERRIBLE_HACK in feature_i.e which
basically amounts to constantly checking if we are currently
processing the default_create of any.

If this isn't there, it won't work at all. I don't know why or how,
but right now it's on the `come-back-to' list.

This commit marks the first version of EVE that can produce some
working (for some degree of `working') SCOOP implementation.