Updated the process_result generator to take into account excluded classes.
The classes which have been excluded from the compilation were previously
still treated as having a SCOOP_SEPARATE implementation which was
clearly false.

This is now remedied by using the SCOOP_BASIC_TYPE module to determine
if this thing should be considered as having a separate interface. This
will have to be updated if we no longer keep a record of which classes
can be separate.