Fixed bug#16598: Melted system execution stops for invariant violation, then resumes by itself
Note: Internally an _invariant routine does not have a proper body index, and this prevent the interp.c to do correctly its job.
For now, we protect from the debugger the access to the stack of _invariant.
But later (after the release), we should fix the issue in the runtime itself.