Fixed for "Evaluator Died" bug when the root class is ANY.default_create.
The bug was introduced when adapting char to EIF_NATIVE_CHAR where we compare
char with EIF_NATIVE_CHAR. By luck, the length of "ANY.default_create" is the
same as "EQA_EVALUATOR.make", and the small loop in eif_init_root for string
comparison did not break until the last character was reached (which is the
same). So "ANY.default_create" was taken as the root while "EQA_EVALUATOR.make"
was expected.

review#11435116.