Changed the way how primary keys are generated.
Instead of creating every key separately, a stored procedure is used
to generate all keys at once. This should reduce the number of round trips
to two on average for any write operation (at least when the metadata about 
the class is already present in the database, which is usually the case).

Also added a new feature last_results in the MySQL and SQLite connectors
to be able to retrieve all results from a multi-statement SQL query.