Removed the generic arguments from PS_COLLECTION_PART and descendants.
The reason for the change was that the current system wasn't flexible enough, since
collections need to inherit from ITERABLE in order to work. This is not the case
for TUPLE however.
The responsibility to iterate over a collection could be moved to the corresponding
COLLECTION_HANDLER, thus eliminating the need for a (constrained) generic parameter 
altogether.