Starting to split OBJECT_DATA.
The idea is to avoid allocating unneeded memory
during read operations for data structures that 
are only used during a write, and vice versa.

This commit makes OBJECT_DATA deferred
and adds two subclasses which can be used for
reading and writing. The ABSTRACT_MANAGER got
a generic argument in order to handle both types
of OBJECT_DATA, and the types in READ_MANAGER,
WRITE_MANAGER and all the handlers are adjusted
to use the new subclasses.

Additionally there's a bug fix for TYPE_METADATA
when dealing with the NONE type.