Performance improvements:
*) Moved {OBJECT_DATA}.references to OBJECT_WRITE_DATA, because
it isn't needed for reading.

*) Removed {OBJECT_DATA}.referers: Complete information about referers
is never used. If in the future a HANDLER wants to map to a relation, it
can only work with a single referer anyway. The only thing it needs to
know is if there are more handlers, to report an error.
Therefore the new mechanism removes the ARRAYED_LIST and introduces
two integers `last_referer' and `referer_count'.

*) Moved {OBJECT_DATA}.to_initialize to OBJECT_READ_DATA, because
it isn't needed for writing.