Changes in the new backend:
Added an individual REPOSITORY implementation for the relational
backend and moved the backend-specific bits to this class.

Changed the return type of {REPOSITORY_CONNECTOR}.stored_types
from LIST [READABLE_STRING_GENERAL] 
to READABLE_INDEXABLE [IMMUTABLE_STRING_8] 

Moved the initialization code which queries the database for
existing types from the factory to the creation procedure.
Rewrote the code to be smarter and only use one SQL query.

Introduced the concept of unmanaged types:
For unmanaged types, ABEL doesn't track the objects identity
and doesn't care about primary keys. It only takes the data
and stores it to the database.
By default all objects used in the new backend are unmanaged.

If a type gets declared as managed, ABEL will handle primary
key generation, and the user is not allowed to touch the primary
key any more.