Added eweasel test#scoop052 that exhibits a bug.
The bug happens when a processor holds a lock
on another processor, and an exception happens.
In that case the lock is not released properly,
which causes an endless wait for the supplier
and a possible deadlock. It also prevents proper
shutdown of the application.

The test is currently marked as manual, because
it never terminates and thus blocks eweasel.