Proper unlocking during an exception.
This commit changes the RTS_SD family of macros
to use the newly introduced functions to keep
track of the request group stack. With this change,
a locked SCOOP processor gets properly unlocked 
when a feature is terminated with an exception.

This commit replaces the old system that used the
thread-local sep_stack and fixes test#scoop052.

See also review#5063692848726016.